requirements.txt
selenium>=4.0.0
undetected-chromedriver>=3.5.0
requests>=2.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
